Moved the different tables into their own Schemas and files
