+++ drizzle/meta/_journal.json
@@ -5,15 +5,8 @@
     {
       "idx": 0,
       "version": "7",
-      "when": 1742409375524,
-      "tag": "0000_lovely_maginty",
-      "breakpoints": true
-    },
-    {
-      "idx": 1,
-      "version": "7",
-      "when": 1742505267779,
-      "tag": "0001_fair_flatman",
+      "when": 1743617776384,
+      "tag": "0000_past_triton",
       "breakpoints": true
     }
   ]
